+++ src/main/java/com/airBnb/AirBnB/service/impl/RoomServiceImpl.java
@@ -4,7 +4,9 @@ import com.airBnb.AirBnB.dto.HotelDto;
 import com.airBnb.AirBnB.dto.RoomDto;
 import com.airBnb.AirBnB.entity.Hotel;
 import com.airBnb.AirBnB.entity.Room;
+import com.airBnb.AirBnB.entity.User;
 import com.airBnb.AirBnB.exception.ResourceNotFoundException;
+import com.airBnb.AirBnB.exception.UnauthorizedException;
 import com.airBnb.AirBnB.repository.HotelRepository;
 import com.airBnb.AirBnB.repository.RoomRepository;
 import com.airBnb.AirBnB.service.InventoryService;
@@ -16,6 +18,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static com.airBnb.AirBnB.util.AppUtils.getCurrentUser;
 
 @Service
 @RequiredArgsConstructor
@@ -31,6 +35,13 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + hotelId));
+
+        User user = getCurrentUser();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnauthorizedException("This user of id "+user.getId()+" is not owner of this hotel");
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -47,6 +58,13 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: " + hotelId));
+
+        User user = getCurrentUser();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnauthorizedException("This user of id "+user.getId()+" is not owner of this hotel");
+        }
+
         return hotel.getRooms()
                 .stream()
                 .map(room -> modelMapper.map(room, RoomDto.class))
@@ -72,6 +90,14 @@ public class RoomServiceImpl implements RoomService {
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with id: " + roomId));
+
+        User user = getCurrentUser();
+
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnauthorizedException("This user of id "+user.getId()+" is not owner of this room");
+        }
+
+
         // Delete future Inventory for this room
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
